// utils/enums.ts
export enum NUMBERS {
  ONE = 1,
  TWO = 2,
  THREE = 3,
  FOUR = 4,
  FIVE = 5,
  SIX = 6,
  SEVEN = 7,
  EIGHT = 8,
}

export enum FORM_FIELDS_NAMES {
  EMAIL = "email",
  PASSWORD = "password",
  CONFIRM_PASSWORD = "confirm_password",
}

export enum BUTTON_COLORS {
  DISABLED = "#eee",
  DEFAULT = "#ffd913",
}

export enum CHARGER_STATUS {
  REMOVED = "REMOVED",
  CONNECTED = "CONNECTED",
  NOT_CONNECTED = "NOT_CONNECTED",
}

export enum CHARGER_TYPE {
  HPC = "HPC",
  "T52" = "T52",
  "T53C" = "T53C",
}